mship/requirements.txt
flask
gunicorn
gevent
gunicorn[gevent]
cassandra-driver
passlib
pyjwt
